requirements.txt
playwright==1.54.0
gspread==6.1.2
google-auth==2.40.3
oauth2client==4.1.3
